(add) continued development of collisions. Overhauled old gjk epa
Contact points might be solved? Although weird that they are so far apart. Followed 2020 paper on collisions and disc/disc collisions work but rects get a lot of spin when colliding. Not sure why. Maybe something to do with ID as they spin in different directions. No idea what the fuck is happening